requirements.txt
Django==2.0
django-allauth==0.37.1
django-appconf==1.0.2
django-celery-results==1.0.1
django-classy-tags==0.8.0
django-compressor==2.1.1
django-cors-headers==2.1.0
django-extensions==1.7.9
django-htmlmin==0.10.0
django-rest-auth==0.9.3
django-sekizai==0.10.0
djangorestframework==3.7.1
drf-nested-routers==0.90.0
psycopg2